feat: add routing for products psges; refactor: change location of some files
