+++ src/components/Burger.module.scss
@@ -1,0 +1,47 @@
+@import '../styles/global';
+
+.burger__button {
+  display: none;
+  background: #d3d3d3;
+  border: #ffffff;
+}
+
+.open {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  a {
+    display: flex;
+    align-items: center;
+    padding: 0 15px;
+    color: $color-dark;
+    font-weight: 700;
+    text-decoration: none;
+    transition: all 0.3s;
+    margin: 10px 0;
+    &.active {
+      color: $color-brown;
+    }
+    &:hover {
+      color: $color-brown;
+    }
+    i {
+      padding: 0 30px;
+    }
+  }
+  @include media-hamburger {
+    .a {
+      display: block;
+    }
+  }
+}
+
+@include media-hamburger {
+  .burger__button {
+    display: block;
+  }
+
+  .navigation {
+    display: none;
+  }
+}
