--- src/components/Layout.module.scss
@@ -1,17 +1,0 @@
-@import '../styles/global';
-
-body {
-  font-family: 'Montserrat', sans-serif;
-  background-color: $color-light;
-  color: $color-dark;
-  margin: 0 auto;
-}
-
-.footer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: $color-general;
-  color: $color-brown;
-  height: 50px;
-}
