+++ src/components/LinksForHeader.tsx
@@ -1,0 +1,26 @@
+import { useContext } from 'react';
+import { AuthContext } from './authProvider';
+import {
+  commonLinks,
+  LinkToLogOut,
+  LinkToRegistration,
+  LinkToSignIn,
+} from './Links';
+
+export const LinksForHeader = (): JSX.Element => {
+  const { isAuth } = useContext(AuthContext);
+
+  return (
+    <>
+      {commonLinks}
+      {isAuth ? (
+        <LinkToLogOut />
+      ) : (
+        <>
+          <LinkToSignIn />
+          <LinkToRegistration />
+        </>
+      )}
+    </>
+  );
+};
