+++ src/index.tsx
@@ -13,10 +13,3 @@ if (container) {
     </BrowserRouter>,
   );
 }
-// getProducts().then(products => console.log('че тут = ', products.body.results));
-// getProductById('df049f34-bcf9-4833-8211-0ba5ddb9870e').then(product =>
-//   console.log('по ИД = ', product),
-// );
-// getProductByKey('body-cream').then(product =>
-//   console.log('по key = ', product),
-// );
