+++ src/pages/ProductPage/ProductPage.module.scss
@@ -1,0 +1,8 @@
+.page {
+  display: flex;
+  justify-content: center;
+  max-width: 1280px;
+  padding: 20px 50px;
+  margin: 0 auto;
+  min-height: calc(100vh - 150px);
+}
